Add files via upload
## Explore the Code
Welcome to "Ghost Killer" - a thrilling Python game built using the Pygame library! Dive into the heart of the action by exploring the codebase on GitHub. 

In this game, you control a courageous player character tasked with defeating enemies and collecting power-ups. With intuitive controls and engaging gameplay, "Ghost Killer" promises an adrenaline-fueled experience for players of all ages.

Feel free to delve into the code, inspect the implementation details, and uncover the secrets behind the game's mechanics.

Embark on a journey through the world of game development and discover the magic of Pygame. Let the adventure begin!